Created a table for all existing products
