+++ shop/src/main/resources/application.properties
@@ -3,5 +3,5 @@ warehouse.baseUrl=https://localhost:8443/warehouse/
 ssl.password=password
 ssl.keystore.location = classpath:warehouse.jks
 
-shop.username=username
+shop.username=avitela
 shop.password=password
